backend/pyproject.toml
[project]
name = "sisprae-utfpr"
version = "0.1.0"
description = ""
authors = [
    {name = "Eduardo Aschwanden Soviersovski",email = "eduardo.as2000@hotmail.com"},
    {name = "Pedro Augusto Tortola",email = "torts@tortsmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.132.0,<0.133.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "authlib (>=1.6.8,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "faker (>=40.11.1,<41.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "aiosmtplib (>=5.1.0,<6.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
